+++ CHANGELOG.md
@@ -7,6 +7,9 @@ This CHANGELOG follows the format listed [here](https://github.com/sensu-plugins
 ### Fixed
 - check-mysql-replication-status: fix code flow if server is not a slave (@DrMurx)
 
+### Changed
+- check-mysql-replication-status: refactoring & spec tests (@DrMurx)
+
 ## [3.1.1] - 2019-03-04
 ### Fixed
 - check-mysql-disk.rb: update defaults for float and integer values as a string as the defaults apparently do not execute the proc which would have properly converted it to the appropriate type (@majormoses)